Updates handling for newer `terraform plan` output
With Terraform v0.10.4, the output of `terraform plan` has changed in two ways
which `terraform-landscape` needed updating:

1. The preface to the plan includes a new header which includes a legend to
explain the actions are creating/updating/removing resources.
2. The output of computed fields has changed to no longer be `<computed>` within
quotes.

This addresses the two updates as follows:

1. Updates `Printer` to scrub for the new preface and skip the
legend. Previously, the legend was being processed as a resource, but didn't
match the resource name pattern.
2. Updates the treetop grammar definition to perform some value sanitization, as
this was the easiest place to centrally handle the permutations it might
encounter.

Added numerous tests to account for forwards and backwards compatibility.

Fixes #30