Fixes an issue with processing attribute names that include colons.
This fixes an issue with processing resource attribute names that included
colons. This is primarily possible when a tag name includes a colon.

The issue is with the parser grammer only matching an attribute name up until
the colon. It was then failing to evaluate the content since it included extra
data. This changes it to use a lookahead matcher to look for `: `.

Note that this still will be a problem if tag names are using `: `, however I'm
unsure if treetop can do a regex that matches the last occurrance within a line.