Sort JSON by keys before generating diff
As discussed in #47 and #57, resource providers return output with keys
in JSON objects sorted arbitrarily.

This can make reading diffs hard even if the net changes are simple.

Since changing the order of keys shouldn't change the interpretation of
the JSON, sort the hashes by key before generating the diff.